feat(routes): move caption routes to blueprint
